Add fetchChunkedLogs. Fetches logs in a while loop. Looping stops on componentWIllUnmount.
